Update authenticationMiddleware.js
Removed the REDIS_PORT from the token blacklist logic as it is no longer necessary. Its reference at the top is fine, as blank will be concatenated on top of the full URL. The current implementation in the imports encapsulates both cases of these env vars being present or not, so no change was needed there.