Update db.js
Currently, if we use ORM functions in a plugin, it will modify the response object, completely hijacking our ability to return custom response messages from a plugin. 

Unsure of whether or not this is a deliberate mechanism, I've added the ability to add a `skipResponse` property to the `options` parameter object, which will simply return the data from the DB ORM Operation instead.